move utils and related tests to folder and split by functionality
